+++ grails-app/services/io/xh/hoist/http/BaseProxyService.groovy
@@ -9,6 +9,7 @@ package io.xh.hoist.http
 
 import groovy.transform.CompileStatic
 import io.xh.hoist.BaseService
+import io.xh.hoist.telemetry.SpanRef
 import org.apache.catalina.connector.ClientAbortException
 import org.apache.hc.core5.http.HttpResponse
 import org.apache.hc.client5.http.entity.UrlEncodedFormEntity
@@ -25,6 +26,9 @@ import org.apache.hc.core5.http.message.BasicNameValuePair
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
+
+import static io.opentelemetry.api.trace.SpanKind.CLIENT
+
 
 @CompileStatic
 abstract class BaseProxyService extends BaseService {
@@ -72,24 +76,40 @@ abstract class BaseProxyService extends BaseService {
         }
         installRequestHeaders(request, method)
 
-        try (CloseableHttpResponse sourceResponse = sourceClient.execute(method)) {
-            response.setStatus(sourceResponse.code)
-            installResponseHeaders(response, sourceResponse)
+        observe()
+            .span(
+                name: request.method,
+                kind: CLIENT,
+                tags: [
+                    'http.request.method': request.method,
+                    'url.full'           : fullPath,
+                    'server.address'     : method.uri.host,
+                    'source'             : 'hoist'
+                ]
+            )
+            .run { SpanRef span ->
+                traceService.injectContext(method)
+                try (CloseableHttpResponse sourceResponse = sourceClient.execute(method)) {
+                    response.setStatus(sourceResponse.code)
+                    span?.setHttpStatus(sourceResponse.code)
+                    installResponseHeaders(response, sourceResponse)
 
-            sourceResponse.entity?.writeTo(response.outputStream)
+                    sourceResponse.entity?.writeTo(response.outputStream)
 
-            response.flushBuffer()
-        } catch (ClientAbortException ignored) {
-            logDebug("Client has aborted request to [$endpoint] - ignoring")
-        } catch (Throwable t) {
-            // Log ...and rethrow exception for normal handling, if not too late
-            logError("Error occurred during proxy streaming of [$endpoint]", t)
-            if (!response.isCommitted()) {
-                response.reset()
-                throw t
+                    response.flushBuffer()
+                } catch (ClientAbortException ignored) {
+                    logDebug("Client has aborted request to [$endpoint] - ignoring")
+                } catch (Throwable t) {
+                    // Log ...and rethrow exception for normal handling, if not too late
+                    logError("Error occurred during proxy streaming of [$endpoint]", t)
+                    if (!response.isCommitted()) {
+                        response.reset()
+                        throw t
+                    }
+                }
             }
-        }
     }
+
 
     //------------------------------------------------
     // Additional overrideable implementation methods