+++ src/main/groovy/io/xh/hoist/cluster/ClusterTask.groovy
@@ -13,6 +13,7 @@ import java.util.concurrent.Callable
 
 import static io.xh.hoist.util.Utils.getIdentityService
 import static io.xh.hoist.util.Utils.getClusterService
+import static io.xh.hoist.util.Utils.getTraceService
 import static io.xh.hoist.util.Utils.getAppContext
 import static io.xh.hoist.json.JSONSerializer.serialize
 
@@ -30,6 +31,8 @@ class ClusterTask implements Callable<ClusterResult>, LogSupport {
     final String username
     final String authUsername
     final boolean asJson
+    final String traceparent
+
 
     ClusterTask(BaseService svc, String method, List args, boolean asJson) {
         this.svc = svc.class.name
@@ -38,13 +41,14 @@ class ClusterTask implements Callable<ClusterResult>, LogSupport {
         this.asJson = asJson
         username = identityService.username
         authUsername = identityService.authUsername
+        traceparent = traceService.captureTraceparent()
     }
 
     ClusterResult call() {
         identityService.threadUsername.set(username)
         identityService.threadAuthUsername.set(authUsername)
 
-        try {
+        try (def traceScope = traceService.restoreContextFromTraceparent(traceparent)) {
             clusterService.ensureRunning()
             def clazz = Class.forName(svc),
                 service = appContext.getBean(clazz),